# perf: use simdutf for idna