@lemire 2025-06-22
Good work. Whether it is worth it depends on the input strings, evidently. :-)