@anonrig 2025-06-29
This is ready to land except:

- updating GitHub workflows to run with simdutf
- add a documentation about this flag